Refactor Tetris.
