+++ Pipfile
@@ -4,8 +4,8 @@ verify_ssl = true
 name = "pypi"
 
 [packages]
+doxa-cli = ">=0.0.2"
 click = ">=8.1.3"
-numpy = ">=1.23.4"
 pillow = ">=9.3.0"
 pygame = ">=2.1.2"
 requests = ">=2.26.0"
