+++ submission/doxa.yaml
@@ -1,4 +1,4 @@
 competition: tetris
 environment: cpu
 language: python
-entrypoint: run.py
+entrypoint: agent.py
