+++ submission/tetris/interface.py
@@ -1,0 +1,3 @@
+class TetrisUI:
+    async def run(self) -> None:
+        raise NotImplementedError
