+++ submission/tetris/pieces/i.py
@@ -1,9 +1,9 @@
-from typing import List, Tuple, Optional
+from typing import List, Optional, Tuple
 
-from engine.piece import TetrisPiece
+from tetris.piece import Piece
 
 
-class IPiece(TetrisPiece):
+class IPiece(Piece):
     piece_type = "I"
 
     def __init__(self) -> None:
