+++ submission/tetris/pieces/l.py
@@ -1,23 +1,23 @@
-from typing import List, Tuple, Optional
+from typing import List, Optional, Tuple
 
-from engine.piece import TetrisPiece
+from tetris.piece import Piece
 
 
-class JPiece(TetrisPiece):
-    piece_type = "J"
+class LPiece(Piece):
+    piece_type = "L"
 
     def __init__(self) -> None:
         super().__init__()
 
-        self.x = 4  # J piece x coordinate is the bottom one in the middle, starts like |_ _ _
+        self.x = 4  # L piece x coordinate is the bottom one in the middle, starts like _ _ _|
         self.y = 2
 
     def spawn_piece(self) -> List[Tuple[int, int]]:
         new = [
-            (self.y - 1, self.x - 1),
             (self.y, self.x - 1),
             (self.y, self.x),
             (self.y, self.x + 1),
+            (self.y - 1, self.x + 1),
         ]
 
         return new
@@ -30,11 +30,11 @@ class JPiece(TetrisPiece):
         if self.orientation == 0:
             if (
                 self.x > 1
-                and not board[self.y - 1][self.x - 2]
                 and not board[self.y][self.x - 2]
+                and not board[self.y - 1][self.x]
             ):
-                old = [(self.y - 1, self.x - 1), (self.y, self.x + 1)]
-                new = [(self.y - 1, self.x - 2), (self.y, self.x - 2)]
+                old = [(self.y, self.x + 1), (self.y - 1, self.x + 1)]
+                new = [(self.y, self.x - 2), (self.y - 1, self.x)]
 
         elif self.orientation == 1:
             if (
@@ -43,7 +43,7 @@ class JPiece(TetrisPiece):
                 and not board[self.y][self.x - 1]
                 and not board[self.y + 1][self.x - 1]
             ):
-                old = [(self.y - 1, self.x + 1), (self.y, self.x), (self.y + 1, self.x)]
+                old = [(self.y - 1, self.x), (self.y, self.x), (self.y + 1, self.x + 1)]
                 new = [
                     (self.y - 1, self.x - 1),
                     (self.y, self.x - 1),
@@ -54,23 +54,23 @@ class JPiece(TetrisPiece):
             if (
                 self.x > 1
                 and not board[self.y][self.x - 2]
-                and not board[self.y + 1][self.x]
+                and not board[self.y + 1][self.x - 2]
             ):
-                old = [(self.y, self.x + 1), (self.y + 1, self.x + 1)]
-                new = [(self.y, self.x - 2), (self.y + 1, self.x)]
+                old = [(self.y + 1, self.x - 1), (self.y, self.x + 1)]
+                new = [(self.y, self.x - 2), (self.y + 1, self.x - 2)]
 
         elif self.orientation == 3:
             if (
                 self.x > 1
-                and not board[self.y - 1][self.x - 1]
+                and not board[self.y - 1][self.x - 2]
                 and not board[self.y][self.x - 1]
-                and not board[self.y + 1][self.x - 2]
+                and not board[self.y + 1][self.x - 1]
             ):
                 old = [(self.y - 1, self.x), (self.y, self.x), (self.y + 1, self.x)]
                 new = [
-                    (self.y - 1, self.x - 1),
+                    (self.y - 1, self.x - 2),
                     (self.y, self.x - 1),
-                    (self.y + 1, self.x - 2),
+                    (self.y + 1, self.x - 1),
                 ]
 
         if old and new:
@@ -86,34 +86,34 @@ class JPiece(TetrisPiece):
         if self.orientation == 0:
             if (
                 self.x < 8
-                and not board[self.y - 1][self.x]
+                and not board[self.y - 1][self.x + 2]
                 and not board[self.y][self.x + 2]
             ):
-                old = [(self.y - 1, self.x - 1), (self.y, self.x - 1)]
-                new = [(self.y - 1, self.x), (self.y, self.x + 2)]
+                old = [(self.y, self.x - 1), (self.y - 1, self.x + 1)]
+                new = [(self.y - 1, self.x + 2), (self.y, self.x + 2)]
 
         elif self.orientation == 1:
             if (
                 self.x < 8
-                and not board[self.y - 1][self.x + 2]
+                and not board[self.y - 1][self.x + 1]
                 and not board[self.y][self.x + 1]
-                and not board[self.y + 1][self.x + 1]
+                and not board[self.y + 1][self.x + 2]
             ):
                 old = [(self.y - 1, self.x), (self.y, self.x), (self.y + 1, self.x)]
                 new = [
-                    (self.y - 1, self.x + 2),
+                    (self.y - 1, self.x + 1),
                     (self.y, self.x + 1),
-                    (self.y + 1, self.x + 1),
+                    (self.y + 1, self.x + 2),
                 ]
 
         elif self.orientation == 2:
             if (
                 self.x < 8
                 and not board[self.y][self.x + 2]
-                and not board[self.y + 1][self.x + 2]
+                and not board[self.y + 1][self.x]
             ):
-                old = [(self.y, self.x - 1), (self.y + 1, self.x + 1)]
-                new = [(self.y, self.x + 1), (self.y + 1, self.x + 2)]
+                old = [(self.y, self.x - 1), (self.y + 1, self.x - 1)]
+                new = [(self.y, self.x + 2), (self.y + 1, self.x)]
 
         elif self.orientation == 3:
             if (
@@ -122,7 +122,7 @@ class JPiece(TetrisPiece):
                 and not board[self.y][self.x + 1]
                 and not board[self.y + 1][self.x + 1]
             ):
-                old = [(self.y - 1, self.x), (self.y, self.x), (self.y + 1, self.x - 1)]
+                old = [(self.y - 1, self.x - 1), (self.y, self.x), (self.y + 1, self.x)]
                 new = [
                     (self.y - 1, self.x + 1),
                     (self.y, self.x + 1),
@@ -142,154 +142,14 @@ class JPiece(TetrisPiece):
         if self.orientation == 0:
             if (
                 self.y < 20
-                and not board[self.y + 1][self.x]
                 and not board[self.y - 1][self.x]
-                and not board[self.y - 1][self.x + 1]
-            ):
-                new = [
-                    (self.y + 1, self.x),
-                    (self.y - 1, self.x),
-                    (self.y - 1, self.x + 1),
-                ]
-
-            elif (
-                self.y == 20
-                and not board[self.y - 1][self.x]
-                and not board[self.y - 2][self.x]
-                and not board[self.y - 2][self.x + 1]
-            ):
-                new = [
-                    (self.y - 1, self.x),
-                    (self.y - 2, self.x),
-                    (self.y - 2, self.x + 1),
-                ]
-
-        elif self.orientation == 1:
-            if self.x == 0:
-                if (
-                    not board[self.y][self.x + 1]
-                    and not board[self.y][self.x + 2]
-                    and not board[self.y + 1][self.x + 2]
-                ):
-                    new = [
-                        (self.y, self.x + 1),
-                        (self.y, self.x + 2),
-                        (self.y + 1, self.x + 2),
-                    ]
-
-            else:
-                if (
-                    not board[self.y][self.x - 1]
-                    and not board[self.y][self.x + 1]
-                    and not board[self.y + 1][self.x + 1]
-                ):
-                    new = [
-                        (self.y, self.x - 1),
-                        (self.y, self.x + 1),
-                        (self.y + 1, self.x + 1),
-                    ]
-
-        elif self.orientation == 2:
-            if (
-                not board[self.y - 1][self.x]
                 and not board[self.y + 1][self.x]
-                and not board[self.y + 1][self.x - 1]
+                and not board[self.y + 1][self.x + 1]
             ):
                 new = [
                     (self.y - 1, self.x),
                     (self.y + 1, self.x),
-                    (self.y + 1, self.x - 1),
-                ]
-
-        elif self.orientation == 3:
-            if self.x == 9:
-                if (
-                    not board[self.y][self.x - 1]
-                    and not board[self.y][self.x - 2]
-                    and not board[self.y - 1][self.x - 2]
-                ):
-                    new = [
-                        (self.y, self.x - 1),
-                        (self.y, self.x - 2),
-                        (self.y - 1, self.x - 2),
-                    ]
-
-            else:
-                if (
-                    not board[self.y][self.x + 1]
-                    and not board[self.y][self.x - 1]
-                    and not board[self.y - 1][self.x - 1]
-                ):
-                    new = [
-                        (self.y, self.x + 1),
-                        (self.y, self.x - 1),
-                        (self.y - 1, self.x - 1),
-                    ]
-
-        else:
-            raise ValueError("Incorrect value for orientation")
-
-        if new:
-            if self.orientation == 0:
-                if self.y == 20:
-                    old = [
-                        (self.y, self.x + 1),
-                        (self.y, self.x - 1),
-                        (self.y - 1, self.x - 1),
-                    ]
-                    self.y -= 1
-
-                else:
-                    old = [
-                        (self.y - 1, self.x - 1),
-                        (self.y, self.x - 1),
-                        (self.y, self.x + 1),
-                    ]
-            elif self.orientation == 1:
-                old = [
-                    (self.y + 1, self.x),
-                    (self.y - 1, self.x),
-                    (self.y - 1, self.x + 1),
-                ]
-                if self.x == 0:
-                    self.x += 1
-            elif self.orientation == 2:
-                old = [
-                    (self.y, self.x - 1),
-                    (self.y, self.x + 1),
                     (self.y + 1, self.x + 1),
-                ]
-            elif self.orientation == 3:
-                old = [
-                    (self.y - 1, self.x),
-                    (self.y + 1, self.x),
-                    (self.y + 1, self.x - 1),
-                ]
-                if self.x == 9:
-                    self.x -= 1
-            else:
-                raise ValueError("Incorrect value for orientation")
-
-            self.orientation = (self.orientation + 1) % 4
-
-        return old, new
-
-    def rotate_anticlockwise(
-        self, board: List[List[Optional[str]]]
-    ) -> Tuple[Optional[List[Tuple[int, int]]], Optional[List[Tuple[int, int]]]]:
-        old, new = None, None
-
-        if self.orientation == 0:
-            if (
-                self.y < 20
-                and not board[self.y - 1][self.x]
-                and not board[self.y + 1][self.x]
-                and not board[self.y + 1][self.x - 1]
-            ):
-                new = [
-                    (self.y - 1, self.x),
-                    (self.y + 1, self.x),
-                    (self.y + 1, self.x - 1),
                 ]
 
             elif (
@@ -306,24 +166,24 @@ class JPiece(TetrisPiece):
 
             else:
                 if (
-                    not board[self.y - 1][self.x - 1]
+                    not board[self.y][self.x + 1]
                     and not board[self.y][self.x - 1]
-                    and not board[self.y][self.x + 1]
+                    and not board[self.y + 1][self.x - 1]
                 ):
                     new = [
-                        (self.y - 1, self.x - 1),
-                        (self.y, self.x - 1),
                         (self.y, self.x + 1),
+                        (self.y, self.x - 1),
+                        (self.y + 1, self.x - 1),
                     ]
 
         elif self.orientation == 2:
             if (
-                not board[self.y - 1][self.x + 1]
+                not board[self.y - 1][self.x - 1]
                 and not board[self.y - 1][self.x]
                 and not board[self.y + 1][self.x]
             ):
                 new = [
-                    (self.y - 1, self.x + 1),
+                    (self.y - 1, self.x - 1),
                     (self.y - 1, self.x),
                     (self.y + 1, self.x),
                 ]
@@ -336,13 +196,13 @@ class JPiece(TetrisPiece):
             else:
                 if (
                     not board[self.y][self.x - 1]
+                    and not board[self.y - 1][self.x + 1]
                     and not board[self.y][self.x + 1]
-                    and not board[self.y + 1][self.x + 1]
                 ):
                     new = [
                         (self.y, self.x - 1),
+                        (self.y - 1, self.x + 1),
                         (self.y, self.x + 1),
-                        (self.y + 1, self.x + 1),
                     ]
 
         else:
@@ -351,41 +211,181 @@ class JPiece(TetrisPiece):
         if new:
             if self.orientation == 0:
                 if self.y == 20:
-                    old = [(self.y, self.x + 1), (self.y - 1, self.x - 1)]
+                    old = [(self.y, self.x - 1), (self.y - 1, self.x + 1)]
                     self.y -= 1
 
                 else:
                     old = [
-                        (self.y - 1, self.x - 1),
                         (self.y, self.x - 1),
                         (self.y, self.x + 1),
+                        (self.y - 1, self.x + 1),
                     ]
             elif self.orientation == 1:
                 if self.x == 0:
-                    old = [(self.y - 1, self.x + 1), (self.y + 1, self.x)]
+                    old = [(self.y - 1, self.x), (self.y + 1, self.x + 1)]
                     self.x += 1
                 else:
                     old = [
-                        (self.y + 1, self.x),
                         (self.y - 1, self.x),
-                        (self.y - 1, self.x + 1),
+                        (self.y + 1, self.x),
+                        (self.y + 1, self.x + 1),
                     ]
             elif self.orientation == 2:
                 old = [
-                    (self.y, self.x - 1),
                     (self.y, self.x + 1),
-                    (self.y + 1, self.x + 1),
+                    (self.y, self.x - 1),
+                    (self.y + 1, self.x - 1),
                 ]
             elif self.orientation == 3:
                 if self.x == 9:
-                    old = [(self.y - 1, self.x), (self.y + 1, self.x - 1)]
+                    old = [(self.y - 1, self.x - 1), (self.y + 1, self.x)]
                     self.x -= 1
                 else:
                     old = [
+                        (self.y - 1, self.x - 1),
                         (self.y - 1, self.x),
                         (self.y + 1, self.x),
+                    ]
+            else:
+                raise ValueError("Incorrect value for orientation")
+
+            self.orientation = (self.orientation + 1) % 4
+
+        return old, new
+
+    def rotate_anticlockwise(
+        self, board: List[List[Optional[str]]]
+    ) -> Tuple[Optional[List[Tuple[int, int]]], Optional[List[Tuple[int, int]]]]:
+        old, new = None, None
+
+        if self.orientation == 0:
+            if (
+                self.y < 20
+                and not board[self.y - 1][self.x - 1]
+                and not board[self.y - 1][self.x]
+                and not board[self.y + 1][self.x]
+            ):
+                new = [
+                    (self.y - 1, self.x - 1),
+                    (self.y - 1, self.x),
+                    (self.y + 1, self.x),
+                ]
+
+            elif (
+                self.y == 20
+                and not board[self.y - 1][self.x]
+                and not board[self.y - 2][self.x]
+                and not board[self.y - 2][self.x - 1]
+            ):
+                new = [
+                    (self.y - 1, self.x),
+                    (self.y - 2, self.x),
+                    (self.y - 2, self.x - 1),
+                ]
+
+        elif self.orientation == 1:
+            if self.x == 0:
+                if (
+                    not board[self.y][self.x + 1]
+                    and not board[self.y][self.x + 2]
+                    and not board[self.y - 1][self.x + 2]
+                ):
+                    new = [
+                        (self.y, self.x + 1),
+                        (self.y, self.x + 2),
+                        (self.y - 1, self.x + 2),
+                    ]
+
+            else:
+                if (
+                    not board[self.y][self.x - 1]
+                    and not board[self.y][self.x + 1]
+                    and not board[self.y - 1][self.x + 1]
+                ):
+                    new = [
+                        (self.y, self.x - 1),
+                        (self.y, self.x + 1),
+                        (self.y - 1, self.x + 1),
+                    ]
+
+        elif self.orientation == 2:
+            if (
+                not board[self.y - 1][self.x]
+                and not board[self.y + 1][self.x]
+                and not board[self.y + 1][self.x + 1]
+            ):
+                new = [
+                    (self.y - 1, self.x),
+                    (self.y + 1, self.x),
+                    (self.y + 1, self.x + 1),
+                ]
+
+        elif self.orientation == 3:
+            if self.x == 9:
+                if (
+                    not board[self.y][self.x - 1]
+                    and not board[self.y][self.x - 2]
+                    and not board[self.y + 1][self.x - 2]
+                ):
+                    new = [
+                        (self.y, self.x - 1),
+                        (self.y, self.x - 2),
+                        (self.y + 1, self.x - 2),
+                    ]
+
+            else:
+                if (
+                    not board[self.y][self.x + 1]
+                    and not board[self.y][self.x - 1]
+                    and not board[self.y + 1][self.x - 1]
+                ):
+                    new = [
+                        (self.y, self.x + 1),
+                        (self.y, self.x - 1),
                         (self.y + 1, self.x - 1),
                     ]
+
+        else:
+            raise ValueError("Incorrect value for orientation")
+
+        if new:
+            if self.orientation == 0:
+                if self.y == 20:
+                    old = [
+                        (self.y, self.x - 1),
+                        (self.y, self.x + 1),
+                        (self.y - 1, self.x + 1),
+                    ]
+                    self.y -= 1
+
+                else:
+                    old = [
+                        (self.y, self.x - 1),
+                        (self.y, self.x + 1),
+                        (self.y - 1, self.x + 1),
+                    ]
+            elif self.orientation == 1:
+                old = [
+                    (self.y - 1, self.x),
+                    (self.y + 1, self.x),
+                    (self.y + 1, self.x + 1),
+                ]
+                if self.x == 0:
+                    self.x += 1
+            elif self.orientation == 2:
+                old = [
+                    (self.y, self.x + 1),
+                    (self.y, self.x - 1),
+                    (self.y + 1, self.x - 1),
+                ]
+            elif self.orientation == 3:
+                old = [
+                    (self.y - 1, self.x - 1),
+                    (self.y - 1, self.x),
+                    (self.y + 1, self.x),
+                ]
+                if self.x == 9:
+                    self.x -= 1
             else:
                 raise ValueError("Incorrect value for orientation")
 
@@ -404,24 +404,24 @@ class JPiece(TetrisPiece):
                 return True
 
         elif self.orientation == 1:
-            if self.y == 19 or board[self.y][self.x + 1] or board[self.y + 2][self.x]:
+            if (
+                self.y == 19
+                or board[self.y + 2][self.x]
+                or board[self.y + 2][self.x + 1]
+            ):
                 return True
 
         elif self.orientation == 2:
             if (
                 self.y == 19
-                or board[self.y + 1][self.x - 1]
+                or board[self.y + 2][self.x - 1]
                 or board[self.y + 1][self.x]
-                or board[self.y + 2][self.x + 1]
+                or board[self.y + 1][self.x + 1]
             ):
                 return True
 
         elif self.orientation == 3:
-            if (
-                self.y == 19
-                or board[self.y + 2][self.x - 1]
-                or board[self.y + 2][self.x]
-            ):
+            if self.y == 19 or board[self.y][self.x - 1] or board[self.y + 2][self.x]:
                 return True
 
         return False
@@ -431,15 +431,15 @@ class JPiece(TetrisPiece):
 
         if self.orientation == 0:
             old = [
-                (self.y - 2, self.x - 1),
+                (self.y - 1, self.x - 1),
                 (self.y - 1, self.x),
-                (self.y - 1, self.x + 1),
+                (self.y - 2, self.x + 1),
             ]
             new = [(self.y, self.x - 1), (self.y, self.x), (self.y, self.x + 1)]
 
         elif self.orientation == 1:
-            old = [(self.y - 2, self.x), (self.y - 2, self.x + 1)]
-            new = [(self.y - 1, self.x + 1), (self.y + 1, self.x)]
+            old = [(self.y - 2, self.x), (self.y, self.x + 1)]
+            new = [(self.y + 1, self.x), (self.y + 1, self.x + 1)]
 
         elif self.orientation == 2:
             old = [
@@ -447,11 +447,11 @@ class JPiece(TetrisPiece):
                 (self.y - 1, self.x),
                 (self.y - 1, self.x + 1),
             ]
-            new = [(self.y, self.x - 1), (self.y, self.x), (self.y + 1, self.x + 1)]
+            new = [(self.y, self.x), (self.y, self.x + 1), (self.y + 1, self.x - 1)]
 
         elif self.orientation == 3:
-            old = [(self.y - 2, self.x), (self.y, self.x - 1)]
-            new = [(self.y + 1, self.x - 1), (self.y + 1, self.x)]
+            old = [(self.y - 2, self.x - 1), (self.y - 2, self.x)]
+            new = [(self.y - 1, self.x - 1), (self.y + 1, self.x)]
 
         else:
             raise ValueError("Incorrect value for orientation")
