+++ submission/tetris/pieces/z.py
@@ -1,9 +1,9 @@
-from typing import List, Tuple, Optional
+from typing import List, Optional, Tuple
 
-from engine.piece import TetrisPiece
+from tetris.piece import Piece
 
 
-class ZPiece(TetrisPiece):
+class ZPiece(Piece):
     piece_type = "Z"
 
     def __init__(self) -> None:
